// src/types/PeakSlice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A window of peaks returned to the display layer.
 */
export type PeakSlice = { samples_per_bucket: number, 
/**
 * Sample position of the first returned bucket.
 */
start_sample: number, 
/**
 * Per channel: interleaved `[min, max, ...]` in i8.
 */
channels: Array<Array<number>>, };